Split MediaWikiServices hook handler
The handler for the hook MediaWikiServices should not has a service
dependency.
